Drop the dtype coercion from TrackedInterval
The aliasing fix only needs interval to be a separate array from topInterval.
Forcing it to float additionally changed behavior for integer bounds, which
the solver never passes, so keep the copy and leave the dtype alone.

+++ yroots/ChebyshevSubdivisionSolver.py
@@ -408,7 +408,7 @@ class TrackedInterval:
         #interval is updated in place by addTransform, so it must not alias topInterval
         #(or the array the caller handed in), otherwise the original interval is lost.
         self.topInterval = interval
-        self.interval = np.array(interval, dtype=float)
+        self.interval = np.array(interval)
         self.transforms = []
         self.ndim = len(self.interval)
         self.empty = False
